refactor: remove old force PhoneInputComponent
