--- src/Components/PhoneNumberInput/index.ts
@@ -1,2 +1,0 @@
-export * from "./PhoneNumberInput"
-export * from "./validatePhoneNumber"
